pyproject.toml
[project]
name = "bo"
version = "0.1.0"
description = ""
authors = [
    {name = "Carlo Bongiovanni",email = "carlo.bongiovanni@ibm.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "panda3d (>=1.10.15,<2.0.0)",
    "pygbag (>=0.9.2,<0.10.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
